Relocate orm.py and data_io to utils #272
